Update patch set 2
Patch Set 2:

Kishen,

This is the start of updates I am planning on doing to update the doxygen documentation for IoTivity-lite.  My plan is to try and do a little documentation work each day to slowly improve the documentation.  limiting the updates to one or two functions per day should make each commit easy to review for accuracy I hope after a few month the API documentation will be vastly improved.

George

Patch-set: 2