Update patch set 3
Patch Set 3:

(6 comments)

From the review it looks to me like you looked mostly at build warnings for your review.

I could change the merge directly but that could result in the change not making it into the swig branch.  I will make the change on the swig branch then update the merge with that change.

Patch-set: 3