Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (6 comments)
> 
> From the review it looks to me like you looked mostly at build warnings for your review.
> 

Well, that is a good place to start. For e.g., I think checking $1.ptr over &$1 provides the intended check at that spot, which we were thus far missing.

I'll look for more leaks in the JNI layer like the one I reported, and, yes, it would be helpful if you too did another pass through the code. It might be good to look in the generated code for a more precise picture.

I'll also be glancing over the Java APIs, but I think for most part they are what they are. However, as I've discussed a few times before, it might be good to add a bit more infrastructure on the Java layer than we currently have, and that could be added after this merge.

> I could change the merge directly but that could result in 
> the change not making it into the swig branch.  I will make 
> the change on the swig branch then update the merge with 
> that change.

Sounds good.

Patch-set: 3