Remove all setter functions from OCResource
This makes it so there are no setter functions for any
of the OCResource properties.

Not some C samples directly modify oc_resource.properties value
it is possible to modify the properties using other functions.
The following should be used instead of setting the OCResource
directly.

OCMain.resourceMakePublic to flip the OC_SECURITY bitmask
OCMain.resourceSetDiscoverable to flip the OC_DISCOVERABLE
bitmask
OCMain.resourceSetDiscoverable and OCMain.resourceSetPeriodicObservable to flip the OC_OBSERVABLE
and OC_PERIODIC bitmask values.

Since the OCResourceTest code relied on the setter functions
for testing the OCResourceTest.java code was deleted.

Change made in responce to reivew comments from merge with
master.

Change-Id: I2f314d19631d9c8bcb31e96e3aee2214c1c2a5c2
Signed-off-by: George Nash <george.nash@intel.com>