Rewrite of swig/README.md based on review
All sections on android were removed except for a pointer
to the port/android/README.md.

All Linux specific commands were moved to a Linux section
instead of being spread around the document.

All Windows specific commands were moved to a Windows section
instead of being spread around the document.

Platform specific parts of the Onboarding and Provisioning
sections were moved to the platform specific sections.

New section added "Windows specific build issues"

The "Verify installation of needed tools" section was moved to
the end of the document since this is mostly a debugging step
it the build has failed.

Small modification to the Layout of swig directory section.

The underline type for sections was changed.  I was using the
the wrong header level in the document.

Other commentary added through out the document.

Change-Id: I88192b228d76e8c4df9d34effe6cb2ebc03adca6
Signed-off-by: George Nash <george.nash@intel.com>