Expose oc_connectivity.h to Java
Additionally added the -Wno-unused-function when building
the swig generated code. Swig generates a lot of code for
dealing with arrays.  Not all of the generated code is
used.  Our choice is to replicate a subsection of the swig
input files or we can simply ignore the warning. Which
is what we are doing.  This is limited only to the swig
generated wrapper code.

This adds the OC_TCP flag when running swig build
so pickup functions that are in a ifdef under OC_TCP.

Change-Id: Idf17fa020c417cb70d726c7e51cfb5112e7c90a6
Signed-off-by: George Nash <george.nash@intel.com>