Updates to address Windows Build issues
Indicate in the README that Rebuild should be used instead
of Build in Visual Studio due to improper detection of build
dependency order.

Add line indicating that JAVA_HOME environment variable must
be set for both Linux and windows.

In the verify installation section I added a check for
JAVA_HOME.

Update build_swig.sh the oc_obt.i file was not building properly
due to the fact it could not find the required include files.

Update the path for build-iotivity-lite.sh to build the jar file
in the iotivity-lite-java/libs directory instead of in the
java_lang directory.

Change-Id: I0c644cf63444eee9b88d0fd3d88a1c1076bd36ef
Signed-off-by: George Nash <george.nash@intel.com>