Add parallel _discovery_all() APIs
These APIs are used to discover all resources on all devices,
and the callback provides an indication of the end of the list
of resources from any discovered device.

This sort of indication could also be achieved with the
pre-existing discovery callback, except that applications
would have to write custom logic to match all resources
per device using (for e.g.) the "anchor" parameter tied
to each discovered resource.

Signed-off-by: Kishen Maloor <kishen.maloor@intel.com>
Change-Id: Iadac84731add2b7f89e02da875197e2a8f1c6581