Update doxygen.ini to increase doc coverage
Add values to the PREDEFINED property. If the values are not
added to the PREDEFINED property then anything #ifdef'ed out
will not be found by doxygen and will not show up in the
generated documentation output.

By adding all the MACROS used to control header file output we
are generating documentation for almost the entier API surface.

Change-Id: I36e06d72bfde183ab329d5286ffb565b42f20a09
Signed-off-by: George Nash <george.nash@intel.com>