Update patch set 4
Patch Set 4:

Hi Nathan!

 > (10 comments)
 > 
 > This looks functionally correct to me, however:
 > 
 > 1) This MUST be pushed to a new branch, not master, as it breaks
 > back-compat with CTT for 2.0.5 and prior.
How can an option to remove something break back-compat? As long as default build stays same - I think there should be no issue with merging it to master.
 > 2) There needs to be a Jira ticket created which points to the BZ
 > entry for this change in the specification.
There is no BZ ticket for this. This was done by request from TSC chair to align with OCF restructuring (e.g. Commercial does not need JW, and JW is optional in spec). Also one consideration is that this decreases dependency on custom modifications in mbedtls.
 > 3) The Jira ticket must be referenced in the commit msg as usual
 > [LITE-nnnn]
I'll create JIRA ticket for this.

Thanks.

Patch-set: 4