Update samples to handle non-integer inputs
When using the Java Scanner to read user input is will
loop if a non-integer value is given. It similar to C scanf
the Java scanner does not clear the input buffer when it fails
to convert a string to an integer. This change updates to code
so it will stop trying to process the same user input if they
give a non-integer value when an integer is expected.

Reproduce issue
run the Java onboarding tool
type 'q' enter. (for some reason I keep typing `q` when I want
to quit the program)

Without this change:
the program will go into a loop that it prints the menu over and
over and does not give the user another chance to enter a new
value.

With this change
It will print "Invalid input. Integer Expected."
then it will wait for user input again.

Change-Id: If4415e3c0b46daf6e713407f2c9c580c60bdf3d1
Signed-off-by: George Nash <george.nash@intel.com>