Expose endpoint member functions in Java
This exposes the endpoint utility functions as member function
of the OCEndpoint class in Java.  Currently the copyList
function is not exposed this way.  The version of the same
functions exposed using OCEndpointUtil class were not removed
as part of this commit but it is likely they will be removed
after review or in a future commit.

Change-Id: I775a4dddd5b1b7acdd69a7151145085eb8c553c9
Signed-off-by: George Nash <george.nash@intel.com>