No longer expose oc_free_server_endpoints
oc_free_server_endpoints depends heavly on the endpoints being a
linked list. Although we don't break the linked lists apart when
passing them to Java but this detail is not transmitted to the
libary user.

This function is problematic and the oc_free_endpoint function
can be used to free all endpoints if needed.

Change-Id: I0b22043710c444f348efc62b45ce93a78c71a9bc
Signed-off-by: George Nash <george.nash@intel.com>