OCF Bridge implementation
This is the initial work to make it possible to implement
an OCF bridge device.

This initial commit gives a function to add the bridge device.
This adds the oic.d.bridge device and add the oic.r.vodlist
resource. As well as a function to respond to GET requests.

Although this is a list to hold the vodlist data there is no
code that will fill the contents of the vodlist yet.

Signed-off-by: George Nash <george.nash@intel.com>