OCF Bridge implementation cont.
This adds the function to add a virtual function.
Unlike the add_bridge function currently is unusable.

This also adds the load_vitural_device_security function
this is copy/paste code from oc_main_init and likely
should be moved to a single location that can be used by
both bridge code and init code.

Signed-off-by: George Nash <george.nash@intel.com>