Fix static analysis warnings from Codan static analysis tool
The codan static analysis tool is built into eclipse CDT which
I used time to time.

Most of these warnings were suggestions to surround bitwise
operations with parenthesis.  Some of the suggetions were
about return values for pthread code in the network_event_thread
function.

In addition a white space error was fixed in the documentation
in oc_api.h

The most common warning is about not break statment at the end
of case statments. All but one of these was done purposly.

Signed-off-by: George Nash <george.nash@intel.com>