Only only list the files with bad formatting whitespace_commit_checker
When originally written the whitespace_commit_checker did not save
a list of the files that had bad formatting. The idea was that
the developer could run clang-format on any of the files and it
would not be a big deal.  However, if the change is large this
results in the tool saying you have x file(s) with bad formating
but then shows every file that was changed. Just showing the files
with bad formatting seems to make more since. It lets the developer
focuse only on the files that need fixing.

Also fixed typo in the output text.

Signed-off-by: George Nash <george.nash@intel.com>