Cleanup for new add device code
Remove debug print statments that were used to develop the
oc_core_add_device_at_index code (e.g. the dump_core_resources
function).

Removed the unreachable code from the oc_core_add_new_device
function. The oc_core_add_new_device function was updated
to call the oc_core_add_new_device_at_index function. This
was done due to the fact that the oc_core_add_device_at_index
and the oc_core_add_new_device code contained almost identical
code with only modifications to handle adding devices at a
given index.

Updated the code for handling when user tries to add a new
device at the same index twice. If the name, spec version,
and data model version are identical to what was already
added the assumption is made that for some reason the add
device code was called twice for the same device and so
the pointer to the existing device is returned. If the strings
do not match then NULL is returned to indicate failure.

Signed-off-by: George Nash <george.nash@intel.com>