Walked though all of the Pojects currently built with Visual Studio and updated the preprocessor defines for each project so they all match the iotivity-lite library build.
The CloudServer was using the old OC_IDD_FILE macro that was
only in the code for a very short while and was eventually
replaced with OC_IDD_API instead.

Anywhere that _CRT_NONSTDC_NO_WARNINGS was used I instead used
_CRT_SECURE_NO_DEPRECATE both of the predefine macros should
have the same effect. In older code (pre-VS2012)the first macro
was used. In newer code the second macro was used. Since the
effect should be the same rather than switch to a differnt macro
for different projects I have chosen to use the newer macro.

Rearanged the order the preprocessor defines are listed in. This
makes no difference on the build but does make a difference when
vewing the macros within the Visual Studio. By placing the
macros that differ between the Debug and Release build last
makes it possible to see most of the macros even when "All
Configurations" option is selected in the properties page.

Signed-off-by: George Nash <george.nash@intel.com>