Overwrite the SWIG generated freeEndpoint code to avoid double free memory error
The OCEndpointUtil.freeEndpoint method will free the memory associated with an
endpoint. If the Java code thinks that it is responsible for freeing the
memory it will also free the endpoint when the GC is run. This results in
freeing memory that was already freed and may be in use by other memory.

This updates the JNI code to check if Java thinks it is responsible for the
memory associated with the endpoint being freews.  If so it will update
the OCEndpoint in question so it no longer thinks its responsible for the
native memory associated with the OCEndpoint. I addition the copy of the
C pointer held in the OCEndpoint class is set to 0 (i.e. NULL) to avoid
having a floating pointer that could cause other problems.

Signed-off-by: George Nash <george.nash@intel.com>