Updated the dummy_bridge to use device info to look up virtual device
This adds code that uses the oc_bridge_get_virtual_device_info function
to look up the virtual device info. A macro was used to make it possible
to use context pointer or the get virtual device info function based on
preference.

Small update to the SCANF macro so it will properly clear the input if
a letter it typed into the input. The fflush should have been sufficent
but it didn't always eat the newline character. It is done with a
empty while loop that calls getchar till EOF is reached.

Signed-off-by: George Nash <george.nash@intel.com>