Updates to oc_bridge code to conform with security spec
The bridge device will not show unowned VODs unless it is
already onboarded.

If the bridge device is reset all unowned VODs will not
be discoverable.

This changes the oc_core_add_device_at_index code so it no
longer automatically calls oc_connectivity_init as soon as the
device is added. The Bridge code takes the responsibility for
handling when oc_connectivity_init and oc_connectivity_shutdown
is called for oic.d.virtual devices.

Discovered bug that caused crash on the ipadapter after
oc_connectivity_shutdown was called and the stack tried to
call oc_send_buffer afterwards.

Signed-off-by: George Nash <george.nash@intel.com>