Update the behavior when the bridge is reset
When the bridge is reset it will cause all of the virtual
devices on that bridge to return to the reset state.

The original intent was to completely remove the the virtual
devices including reseting the VOD map using the new
`oc_vod_map_reset` function. This function can not be used till
code is introduced that will release the virtual device.
Then the VOD map can be reset. The VOD map is required for
proper operation of virtual devices.

Removed the change to ipadapter since with this change dev
pointer is no longer change to NULL while processing a callback.

Fixed the oc_vod_map_free function to properly clear out the
VODs from the vod_list as well as to set the pointer NULL after
its freed.

Signed-off-by: George Nash <george.nash@intel.com>