Add oc_remove_virtual_device function
This will remove the virtual device without reseting any SVR for the
device.

Null check added to oc_connectivity_shutdown. If shutdown is called
when the connectivity is already shutdown then the ip_context will
be Null.  This will make it so calling shutdown on an already
shutdown connection will not cause a segfault.

Signed-off-by: George Nash <george.nash@intel.com>