Update dummy_bridge to enable simpler testing
The dummy bridge can now simulate removing a virtual device.

Added ability to reset bridge and the virtual devices using
the UI.

Updated the code that created the directory to use the
`CreateDirectory` function.  The `mkdir` function worked
on Windows but the compiler complained about it being
an undefined function at compile time.

Fixed build failure when built with SECURITY=0

Signed-off-by: George Nash <george.nash@intel.com>