Fix bug when removing ownership status callback from callback
A developer may decide to remove the oc_ownership_status_cb_t
from inside the callback hander code. The issue is that
removing the callback invalidates that pointer for the list
that holds the oc_ownership_status_cb_t. Causing the code
that walks that list to fail when trying to get the next
item from the list. This change will make it so the invoking
code changes to the next item before the callback is invoked.

Signed-off-by: George Nash <george.nash@intel.com>