Add APIs to Java
This exposes oc_endpoint_string_parse_path to Java
under OCEndpointUtil.endpointStringParsePath

Ignore oc_ri_is_client_cb_valid this function is not
expected to be used in Java.

With oc_endpoint_string_parse_path exposed to Java we are
able to update the OCEndpointUtil.stringToEndpoint method so
it can take only one parameter. Since developers are already
using the stringToEndpoint method that takes two parameters
we need to still provide that method.

The stringToEndpoint that takes two parameters has a return
value in one of the parameters.  Although this works it is
very unsual in Java to use code this way. Now users can call
the single parameter version of stringToEndpoint and if they
need the path from that endpoint they can call
endpointStringParsePath using the same endpoint string. More
method call but much closer to what a Java developer would
expect.

Signed-off-by: George Nash <george.nash@intel.com>