Fix shutdown failure when using virtual devices
When using virtual devices the first time a device is started
VOD1 will get assigned an index (1) then VOD2 will get another
index (2). If the device is shut down and restarted then only
VOD2 is started. It will be assigned the same index (2) so that
its used the same security files. This result is that the acl and creds are not initilized for VOD1.  This will cause a
segfault when the shutdown code tries to clear the acl and creds
for the device that was not initilized.

Before clearning the acl and creds we check that the device was
initilized by checking that the di is not a nil UUID. If not
initilized the function oc_sec_clear_acl and oc_sec_clear_creds
return instantly.

Signed-off-by: George Nash <george.nash@intel.com>