Helper functions added to oc_uuid
This adds the two new helper functions to oc_uuid
- oc_uuid_is_equal_to: everywhere in the code where two uuids
  are compaired a simple memcmp is used but that relies on knowing
  the size of the uuid is 16 bytes this abstracts away the need to
  know that fact.
- oc_uuid_copy: make a copy of the uuid once again this could be done
  really quickly using memcpy but this reduces the knowledge required
  to use the uuid.

Fixed issue found when running the memory sanitizer and the unit test
StrToUUIDTest_P the test the sanitizer showed that the oc_str_to_uuid
would run past the end of a uuid for the supplied uuid.

In the test was checking if the oc_str_to_uuid would parse a uuid
with out '-' characters. It would but it also would run past the
end of the string. This updates to code to prevent running past
the end of the sring and expanded the test a little more to
verify it was behaving as expected.

Signed-off-by: George Nash <george.nash@intel.com>