Fix memory sanitizer issues
Fix crash in ipadapter if oc_connectivity_shutdown was
called for an already shutdown connection.

Fix memory issue that the oc_vods_list_t was pointing to an
incorrect value due to a realloc of memory. Instead of
pointing to the device id it is instead copied into the vodslist

Signed-off-by: George Nash <george.nash@intel.com>