Clear piid on delete virtual device
When a virtual device is deleted the piid is set to nil uuid.

The add_virtual_device code now checks the piid to see if
it is nil.  If is it nil a new random uuid will be generated.

Small update to the dummy_bridge sample to make handle the
delete virtual device code.

Signed-off-by: George Nash <george.nash@intel.com>