cloud: do not try to reconnect on selected errors
When refresh_token or register requestes receive
SERVICE_UNAVAILABLE or GATEWAY_TIMEOUT responses
try to reconnect to cloud; on other errors the
cloud closes connection and does not try
reconnecting.
Additionally, when this happens for refresh_token
change the state of the device and remove logged in
and registered flags.