cloud: correctly delete created resources on reset
On factory reset dynamically created resources are deleted by
oc_rt_factory_free_created_resources. The function is however
called after device id has been reset to a new random value.
Deletion is executed only after a resource is unpublished from
the cloud, but the unpublish requests never suceed because
the cloud does not recognize the new device ID and thus
the created devices remain in to-be-deleted state forever.
The issue is fixed by not generating unpublish request during
factory reset and proceeding to the deletion directly.