collection: keep links in ordered by uri
Each collections keeps a list of owned resources (links).
Default uri for a resource in a collection is in the format
"${collection uri}/${index}", where index is a unique
integer value.

When a new resource is generated a unique index value
must be found. We also want index to be the lowest
possible value. For an unordered list this would require
an O(n^2) algorithm.
To enable the use of an O(n) algorithm we keep the list
with links ordered primarily by uri length and secondarily
by uri value.