Fix issue oic/d piid issue with notification (#40)
When generating the representation for the /oic/d
resource the getter checks the endpoint_t version
provided as argument to the call. If the endpoint
is NULL or has an old version the piid field
is skipped.
When sending observation messages NULL was used
as the endpoint value, therefore the piid field
was never present in the notifications.

The issue is fixed by taking the endpoint value
from the currently handled observer.