security: fix trust anchors chain
Iotivity-lite contains two containers that store trust anchors.
One is a simple linked list and second is a mbedtls chain. When
a trust anchor was deleted it was correctly removed from the
linked list, but the mbdetls was handled incorrectly. The chain
is now cleared and reloaded correctly from the linked list.