Update SonarCloud workflow
sonnar-scanner currently doesn't handle multi configuration
compilation. Thus if a file is compiled multiple times (such is
the case for files in the api folder) then only the first
compilation of such a file is analysed. So the best we can do
is enable as many features as possible and compile target in
such an order to get most of the code analysis (ie. compile
client-server target first, so we get analysis with both
OC_CLIENT and OC_SERVER defined)