Install mbedtls also when used as static library (#346)
* Correct installation must include installation of static libraries which are used by
iotivity-lite libraries. Iotivity-lite static library targets depend on mbedtls static libraries, 
and CMake propagates this dependency when one is creating an executable which
links with an iotivity-lite static library. Therefore mbedtls libraries must be included 
in the iotivity-lite installation.

* Add OC_INSTALL_MBEDTLS CMake option (enabled by default) which controls
whether mbedtls is included in the iotivity-lite installation. Set it to OFF to not 
include mbedtls in installation.