Use public API to get certificate from SSL config
Previously an internal mbedTLS header was imported which
exposed additional structures, which are not available in
the public API. Such an internal structure was used to get
SSL certificate from the SSL configuration.
We have switched to use mbedtls_ssl_conf_iterate_own_certs
and removed the import of the internal header.