cloud: use timeout to schedule retries
In previous implementation cloud_register, cloud_login and
cloud_refresh_token handled retry in the following way:
  - when executing a call (cloud_login, ...) also schedule another 
execution after a retry interval
  - in the response handler, remove the scheduled retry

When the server is slow to respond this might result in the
response handler being invoked multiple times. For cloud manager,
that always ends in a failure because multiple invocations
of the same handler result in an invalid state.

The reworked version doesn't schedule retry when executing a call.
Instead waiting for response ends with timeout error
(code OC_STATUS_SERVICE_UNAVAILABLE) if the server doesn't respond
in time. When timeout is detected only then is retry scheduled.