cloud: Fix heap-use-after-free when deregisting from cloud
Setting empty server forces deregistration from the original
cloud server (if the device is connected to it). This is done
by sending a deregister request to the cloud.
When the deregistration is triggered by a receiving a POST request
form the server then creating a new request ovewrites the global
shared buffer for requests. Because the original POST request is
waiting for response and will try to use the buffer already used
and freed by the deregistration request.

The solution is to execute deregistration only after the POST
response has been sent.