Validate incoming messages in network thread
When invalid data comes from TCP connection, the connection will
be dropped. The network thread will now perform a validation check
on incoming data via TCP. Any data that fails the validation check
will result in the termination of the connection.

For UDP, an invalid incoming message is dropped.

Fixes #417