PlgdTime: Fetch time from endpoint (#408)
* plgd/time: fetch time from endpoint

Add plgd_time_fetch to plgd_time feature to retrieve time from
an endpoint.

For TLS/DTLS if a session to the endpoint is not yet
established then the feature tries to create a new session.
However, since the device might have a desynchronized time
the certificate verification functions are overriden to
ignore time validity. This might be a security risk, so once
a response is received this connection is scheduled to be
closed.
The client might be required to wait for the session to
disconnect before reconnecting to the endpoint. This can be
achieved using the sessions_events API.