Fix versioning scheme in the Publish docker images job
Switch to pep440 versioning scheme in the helper Github action
docker/metadata-action@v4. It seems that the IoTivity versioning
format "X.Y.Z.R" (where "X.Y.Z" is the OCF Specification Version
and "R" is the version of the C library) is no longer recognized
by the action as matching the format of semantic versioning.