Fix deletion of a resource by a CoAP request
After a DELETE request handler returns some processing from
the coap engine is still done with the resource, therefore
the resource cannot be deleted directly by oc_ri_delete_resource.
Instead it must be scheduled for deletion by
oc_delayed_delete_resource.
However, a resource scheduled for deletion must not be used once
the processing is done. Therefore it is removed from the list of
application resources and added to the list of resources scheduled
to be deleted.
To check whether your resource is scheduled to be deleted you can
use oc_ri_is_app_resource_to_be_deleted.