Linux: Configure listening ports of a device
Linux port now provides a configuration option to specify port
numbers or disable ports for each IP protocol. The following are
the effects of disallowing specific protocols:

- Disabling UDP4 or UDP: The device will no longer be discoverable
via multicast for certain protocol, and the connection will not work
as both rely on the same socket.
- Disabling DTLS4 or DTLS: DTLS will be disabled for both the server
and the client.
Disabling TCP4 or TCP or TLS(TCP)4 or TLS(TCP): Only the TCP listener
will be disabled, while the client side will continue to function.

Example: apps/cloud_server.c