security:  when a connection has been closed, drop the messages associated with it (#446)
* security: when a connection has been closed, drop the messages associated with it

Fixes #441

After the TCP-TLS/DLTS session is closed, there may be queued messages in the TLS queue that are still being processed. If this occurs, the system will attempt to create a new connection. However, this can fail if the TLS peer is not properly configured, such as having the wrong role (server instead of client) or not allowing the appropriate ciphers for cloud connections.

Therefore, it is necessary to discard all messages associated with the events queue when the session ends.