Fix runtime disabling of logging
Updated the internal representation from an unsigned to a signed
integer. The value OC_LOG_LEVEL_DISABLED (=-1) is now to disable
logging during runtime. Whether a log level is enabled is checked
by the '<' operator. Previously, when casting to uint8_t, the value
overflowed to 255, causing all log levels to be enabled.
This issue has been resolved by using a signed integer representation.