android: disable network interface monitoring on ANDROID_API >= 30
The current implementation of network interface changes notifications
is based on binding to a NETLINK socket and using select to monitor
changes. However, from ANDROID_API=30 the binding to a NETLINK socket
is restricted and the bind call will fail, which ultimately causes
the device to abort during initialization.
This commit simply disables the feature when ANDROID_API>=30 is
detected.

For ANDROID_API>=30 this feature needs to be reimplemented using a
non-restricted API.