Implement Entity Tag (ETag) for resources
This commit introduces the Entity Tag (ETag) feature, enabling
resources to have their own ETag values. The ETag value of
a resource can be updated by calling either oc_resource_update_etag
or oc_notify_resource_changed, the latter also notifying observers
if the resource is observable.

If a resource supports POST or PUT requests, its ETag is
automatically updated upon a successful request.

The value of the ETag is based on the system time, using
oc_clock_time(). If OC_STORAGE is enabled, the current ETag can be
stored and loaded from storage. For more details on the public API,
refer to the etag.h header file.

In CoAP responses to GET requests addressed to a specific resource,
the ETag is included in CoAP options if the response code is
Content (2.05) or Valid (2.03). If a GET request contains an ETag
in CoAP options that matches the ETag of the addressed resource,
the response code will be Valid (2.03), and no data will be
included in the response.

These changes enhance the efficiency of resource handling in the
IoTivity-lite library.