ETag for GET Requests on /oic/res with batch interface
When a GET request is made to the /oic/res resource with the batch
interface, an array of contents representing discoverable and
accessible resources is returned. This enhancement focuses on
the calculation and handling of ETags in this scenario.

ETags are utilized to track changes in resources. The ETag for each
resource is updated whenever a change occurs. In the current IoT
implementation the ETag value for resources continually increases,
ensuring uniqueness among resources.

In the context of the batch interface for the /oic/res resource
the ETag value is calculated. The highest ETag value is selected
from the resources included in the payload as well as the /oic/res
resource itself as the ETag of the payload. This approach
guarantees that the ETag used for the batch interface accurately
represents the state of the included resources.