Implement Incremental Updates for /oic/res Resource
This update introduces the Incremental Updates feature for the
/oic/res resource in CoAP requests. When querying the /oic/res
resource using a GET request with the batch interface and including
the "incChanges" key in the query, Incremental Updates are
activated. The "incChanges" key should contain a comma-separated
list of base64 encoded ETags (using URL-safe encoding without
padding).

This feature leverages the uniqueness and strict incrementality
of ETags. By comparing the ETags received in the query with the
stored ETags, the system identifies the highest still valid ETag.
Resources with an ETag that is lower or equal than the identified
ETag are skipped and not contained by the GET request response,
since we know that they have not been updated.