ETag: Add CRC64 checksum for each stored resource
- Implemented the calculation of a CRC64 checksum for each resource
payload.
- By default, the GET payload with the "oic.if.baseline" interface
is used for a resource.
- Special handling is added for resources `/x.plgd.dev/time` and
`/oic/res`, which contain properties that might change on each GET
request. For these resources, an internal "x.plgd.if.etag"
interface is used to filter out problematic properties from the
payload.
- Updated the ETag storage payload to include not only the ETag
of the given resource but also the checksum for the resource.
- When data is loaded from storage, it now verifies the checksum
before processing, ensuring data integrity.