Implement Message Queue Limit to Prevent Memory Exhaustion in Embedded Devices
The current network thread can produce unlimited messages to the queue
which is consumed by the event thread. In embedded devices, we have limited
memory, so a DDOS attack could exhaust the memory.

Fixes Issue: #626

To fix this issue, (when OC_DYNAMIC_ALLOCATION is enabled) we use the macro
OC_MAX_NUM_CONCURRENT_REQUESTS , which defines the maximum number
of messages in the queue per device.
The default value of OC_MAX_NUM_CONCURRENT_REQUESTS is 32.

If the number of messages reaches the limit, all file descriptors in the network
thread that produce oc_message_t are removed from select/poll. The file
descriptor for monitoring the control flow is always set and is used to wake up
from select() when the queue is ready to accept events again.
When more file descriptors are ready to be read than the number of available
queue slots, a subset is randomly selected to be pushed to the queue.

---------

Co-authored-by: sp-milos <bazelides@siprogs.sk>
Co-authored-by: Daniel Adam <daniel.adam1922@protonmail.com>